# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Koumbaya Marketplace API",
        "description": "API REST pour la marketplace Koumbaya avec système de tombolas et paiements mobile money",
        "contact": {
            "name": "Équipe Koumbaya",
            "email": "dev@koumbaya.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement Koumbaya"
        }
    ],
    "paths": {
...
        "/api/otp/send": {
            "post": {
                "tags": [
                    "OTP"
                ],
                "summary": "Envoyer un code OTP",
                "operationId": "d34d0fcd867e74d90714f6163dae0c32",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "identifier",
                                    "type",
                                    "purpose"
                                ],
                                "properties": {
                                    "identifier": {
                                        "description": "Email ou numéro de téléphone",
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "type": {
                                        "description": "Type d'envoi",
                                        "type": "string",
                                        "enum": [
                                            "email",
                                            "sms"
                                        ],
                                        "example": "email"
                                    },
                                    "purpose": {
                                        "description": "But du code OTP",
                                        "type": "string",
                                        "enum": [
                                            "registration",
                                            "password_reset",
                                            "login",
                                            "payment"
                                        ],
                                        "example": "registration"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Code OTP envoyé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Code de vérification envoyé par email"
                                        },
                                        "data": {
                                            "properties": {
                                                "masked_identifier": {
                                                    "type": "string",
                                                    "example": "us**@example.com"
                                                },
                                                "expires_at": {
                                                    "type": "string",
                                                    "example": "2025-08-07 18:05:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Trop de tentatives",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Trop de tentatives, veuillez patienter"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/otp/verify": {
            "post": {
                "tags": [
                    "OTP"
                ],
                "summary": "Vérifier un code OTP",
                "operationId": "4d68b7909e86f7183a88d99c570feee3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "identifier",
                                    "code",
                                    "purpose"
                                ],
                                "properties": {
                                    "identifier": {
                                        "description": "Email ou numéro de téléphone",
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "code": {
                                        "description": "Code OTP à 6 chiffres",
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "purpose": {
                                        "description": "But du code OTP",
                                        "type": "string",
                                        "enum": [
                                            "registration",
                                            "password_reset",
                                            "login",
                                            "payment"
                                        ],
                                        "example": "registration"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Code OTP vérifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Code de vérification validé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Code invalide ou expiré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Code de vérification invalide ou expiré"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/otp/resend": {
            "post": {
                "tags": [
                    "OTP"
                ],
                "summary": "Renvoyer un code OTP",
                "operationId": "fd23a34799a79abfe400251392bba317",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "identifier",
                                    "type",
                                    "purpose"
                                ],
                                "properties": {
                                    "identifier": {
                                        "description": "Email ou numéro de téléphone",
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "type": {
                                        "description": "Type d'envoi",
                                        "type": "string",
                                        "enum": [
                                            "email",
                                            "sms"
                                        ],
                                        "example": "email"
                                    },
                                    "purpose": {
                                        "description": "But du code OTP",
                                        "type": "string",
                                        "enum": [
                                            "registration",
                                            "password_reset",
                                            "login",
                                            "payment"
                                        ],
                                        "example": "registration"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Code OTP renvoyé avec succès"
                    }
                }
            }
        },
        "/api/otp/status/{identifier}": {
            "get": {
                "tags": [
                    "OTP"
                ],
                "summary": "Vérifier le statut d'un OTP",
                "operationId": "49bd7d71599eff2916f33c8aa045dcde",
                "parameters": [
                    {
                        "name": "identifier",
                        "in": "path",
                        "description": "Email ou téléphone",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "purpose",
                        "in": "query",
                        "description": "But du code OTP",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "registration",
                                "password_reset",
                                "login",
                                "payment"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statut de l'OTP",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "has_pending_otp": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "masked_identifier": {
                                                    "type": "string",
                                                    "example": "us**@example.com"
                                                },
                                                "expires_at": {
                                                    "type": "string",
                                                    "example": "2025-08-07 18:05:00"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "email"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/otp/cleanup": {
            "delete": {
                "tags": [
                    "OTP"
                ],
                "summary": "Nettoyer les anciens codes OTP (Admin only)",
                "operationId": "32b1e92d3fb6208119c7a75ed92e37d8",
                "responses": {
                    "200": {
                        "description": "Nettoyage effectué",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Nettoyage effectué"
                                        },
                                        "data": {
                                            "properties": {
                                                "deleted_count": {
                                                    "type": "integer",
                                                    "example": 15
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification via token Bearer JWT",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}